rename to shwap
