+++ share/shwap/sample_hasher.go
@@ -1,4 +1,4 @@
-package ipldv2
+package shwap
 
 import (
 	"crypto/sha256"
