+++ share/shwap/sample_id.go
@@ -1,4 +1,4 @@
-package ipldv2
+package shwap
 
 import (
 	"encoding/binary"
